fixed code review comments
