--- common/constructs/lib/ec2/index.ts
@@ -1,1 +1,0 @@
-export * from './instance';
