+++ initial-setup/templates/src/deployments/account-warming/index.ts
